Scheduler moved into the common library. It's 'TaskSchduler' now and can be shared between services that need it
