+++ common/src/datasources/datasources.ts
@@ -1,9 +1,9 @@
 import { MarketDataPoint } from "../models/marketDataPoint";
 import { IndicatorDataPoint } from "../models/indicatorDataPoint";
 
-// Constrain T so it’s either a MarketDataPoint or an IndicatorDataPoint
-export interface DataSource<T extends MarketDataPoint | IndicatorDataPoint> {
+export type TaskResult = MarketDataPoint | IndicatorDataPoint | (MarketDataPoint | IndicatorDataPoint)[];
+
+export interface DataSource {
 	key: string;
-	// now returns an array of exactly T
-	fetch(): Promise<T[] | null>;
+	fetch(): Promise<TaskResult | null>;
 }
