+++ common/src/models/data.ts
@@ -1,3 +1,4 @@
+import { DataSource } from "../datasources/datasources";
 import { IndicatorDataPoint } from "./indicatorDataPoint";
 import { MarketDataPoint } from "./marketDataPoint";
 
@@ -6,3 +7,15 @@ export type DataSourceResult = MarketDataPoint | IndicatorDataPoint | (MarketDat
 
 // union type for supported message types
 export type SupportedMessage = MarketDataPoint | IndicatorDataPoint;
+
+export interface ScheduledDataSource {
+	source: DataSource;
+	schedule: string; // cron expression
+	enabled: boolean;
+	priority: "high" | "medium" | "low";
+	maxRetries: number;
+	retryDelay: number; // milliseconds
+	lastRun?: Date;
+	lastSuccess?: Date;
+	consecutiveFailures: number;
+}
