+++ data-collector/src/datasources/feargreed.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 import { DataSource } from "@financialsignalsgatheringsystem/common";
 import { IndicatorDataPoint, Signal } from "@financialsignalsgatheringsystem/common";
 
-export class FearGreedSource implements DataSource<IndicatorDataPoint> {
+export class FearGreedSource implements DataSource {
 	public key = "fear_greed_index";
 
 	async fetch(): Promise<IndicatorDataPoint[] | null> {
