+++ data-collector/src/datasources/fred.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 import { DataSource } from "@financialsignalsgatheringsystem/common";
 import { IndicatorDataPoint } from "@financialsignalsgatheringsystem/common";
 
-export class FredSource implements DataSource<IndicatorDataPoint> {
+export class FredSource implements DataSource {
 	public key: string;
 	private readonly apiKey: string;
 	private readonly series_id: string;
