+++ package.json
@@ -4,7 +4,7 @@
   "version": "1.0.0",
   "description": "",
   "workspaces": [
-    "common", 
+    "common",
     "data-collector",
     "signal-persister",
     "signal-query-api",
@@ -26,13 +26,16 @@
     "@types/node": "^22.15.19",
     "@typescript-eslint/eslint-plugin": "^8.32.1",
     "@typescript-eslint/parser": "^8.32.1",
-    "eslint": "^9.27.0",
-    "eslint-config-prettier": "^10.1.5",
-    "eslint-plugin-prettier": "^5.4.0",
-    "globals": "^16.1.0",
+    "eslint": "^9.31.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.5.3",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "globals": "^16.3.0",
     "nodemon": "^3.1.10",
     "prettier": "^3.5.3",
-    "typescript-eslint": "^8.32.1"
+    "typescript-eslint": "^8.37.0"
   },
   "dependencies": {
     "axios": "^1.9.0",
