+++ scraper-service/src/index.ts
@@ -1,34 +1,27 @@
 import { RabbitMQService } from "@financialsignalsgatheringsystem/common";
+import { ScraperConfigManager } from "./scheduler/ScraperConfigManager";
 import { config } from "./utils/config";
-import { ScrapeScheduler } from "./scheduler/ScrapeScheduler";
-
-import { AppStoreRankScraper } from "./scrapers/AppStoreRankScraper";
-
 class ScraperApp {
 	private messageBroker: RabbitMQService;
-	private scheduler: ScrapeScheduler;
+	private scraperManager: ScraperConfigManager;
+	// ... other services
 
 	constructor() {
 		this.messageBroker = new RabbitMQService(config.RABBITMQ_URL!);
-		this.scheduler = new ScrapeScheduler(this.messageBroker);
+		this.scraperManager = new ScraperConfigManager(this.messageBroker);
+		// ... instantiate other services
 	}
 
 	public async start(): Promise<void> {
-		console.log("[ScraperApp] Starting scraper service...");
 		await this.messageBroker.connect();
-		console.log("[ScraperApp] Connected to RabbitMQ.");
-
-		this.scheduler.schedule(
-			new AppStoreRankScraper("coinbase", "id886427730", "us"),
-			"0 8 * * *" // Run once a day at 8:00 AM UTC
-		);
-
-		console.log("[ScraperApp] All scrapers have been scheduled.");
+		this.scraperManager.setupDefaultSchedules();
+		this.scraperManager.getScheduler().start();
+		console.log("[ScraperApp] Scraper service started and tasks scheduled.");
+		// ... start API, monitor, etc.
 	}
+
+	// ... graceful shutdown
 }
 
 const app = new ScraperApp();
-app.start().catch((err) => {
-	console.error("[ScraperApp] Fatal error during startup:", err);
-	process.exit(1);
-});
+app.start();
