+++ scraper-service/src/models/Scraper.interface.ts
@@ -5,5 +5,5 @@ export type ScraperResult = IndicatorDataPoint | MarketDataPoint[] | null;
 export interface Scraper {
 	key: string;
 	// The main execution method
-	scrape(): Promise<MarketDataPoint[] | null>;
+	fetch(): Promise<MarketDataPoint[] | null>;
 }
