+++ scraper-service/src/utils/config.ts
@@ -19,5 +19,4 @@ export const config = {
 	HEALTH_HEAP_WARNING_MB: parseInt(process.env.HEALTH_HEAP_WARNING_MB || "512", 10),
 	HEALTH_HEAP_CRITICAL_MB: parseInt(process.env.HEALTH_HEAP_CRITICAL_MB || "768", 10),
 	MONITOR_CHECK_INTERVAL_CRON: process.env.MONITOR_CHECK_INTERVAL_CRON || "*/5 * * * *",
-	API_NINJAS_KEY: getEnv("API_NINJAS_KEY"),
 };
