+++ signal-persister/src/services/database.interface.ts
@@ -1,8 +1,7 @@
-import { Signal } from "@financialsignalsgatheringsystem/common";
 import { MarketDataPoint, IndicatorDataPoint } from "@financialsignalsgatheringsystem/common";
 export interface DatabaseService {
 	connect?(): Promise<void>; // Optional if connection is managed by pool instantiation
-	insertSignal?(signal: Signal): Promise<void>;
+
 	insertIndicator?(point: IndicatorDataPoint): Promise<void>;
 	insertMarketData?(point: MarketDataPoint): Promise<void>;
 	// getSignalsByName(name: string, limit: number): Promise<Signal[]>; // For a potential API service
