+++ tsconfig.base.json
@@ -5,6 +5,7 @@
       "@financialsignalsgatheringsystem/common": ["common/src"], 
       "@financialsignalsgatheringsystem/common/*": ["common/src/*"]
     },
+    "rootDir": ".",
     "target": "es2021",
     "module": "commonjs",
     "lib": ["es2021"],
